Moved unit test from core unit tests to stand-alone unit test in producer plugin
