+++ plugins/producer_plugin/CMakeLists.txt
@@ -9,3 +9,5 @@ add_library( producer_plugin
 target_link_libraries( producer_plugin chain_plugin signature_provider_plugin appbase eosio_chain )
 target_include_directories( producer_plugin
                             PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/include" "${CMAKE_CURRENT_SOURCE_DIR}/../chain_interface/include" )
+
+add_subdirectory( test )
