+++ plugins/producer_plugin/test/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_executable( test_snapshot_information test_snapshot_information.cpp )
+target_link_libraries( test_snapshot_information producer_plugin eosio_testing )
+
+add_test(NAME test_snapshot_information COMMAND plugins/producer_plugin/test/test_snapshot_information WORKING_DIRECTORY ${CMAKE_BINARY_DIR})
+
