+++ unittests/snapshot_tests.cpp
@@ -5,8 +5,7 @@
 #include <eosio/chain/kv_chainbase_objects.hpp>
 #include <eosio/chain/snapshot.hpp>
 #include <eosio/testing/tester.hpp>
-#include <eosio/producer_plugin/producer_plugin.hpp>
-#include <eosio/producer_plugin/pending_snapshot.hpp>
+#include <eosio/testing/snapshot_suites.hpp>
 
 #include <boost/mpl/list.hpp>
 #include <boost/test/unit_test.hpp>
@@ -79,109 +78,7 @@ public:
    bool validate() { return true; }
 };
 
-struct variant_snapshot_suite {
-   using writer_t = variant_snapshot_writer;
-   using reader_t = variant_snapshot_reader;
-   using write_storage_t = fc::mutable_variant_object;
-   using snapshot_t = fc::variant;
-
-   struct writer : public writer_t {
-      writer( const std::shared_ptr<write_storage_t>& storage )
-      :writer_t(*storage)
-      ,storage(storage)
-      {
-
-      }
-
-      std::shared_ptr<write_storage_t> storage;
-   };
-
-   struct reader : public reader_t {
-      explicit reader(const snapshot_t& storage)
-      :reader_t(storage)
-      {}
-   };
-
-
-   static auto get_writer() {
-      return std::make_shared<writer>(std::make_shared<write_storage_t>());
-   }
-
-   static auto finalize(const std::shared_ptr<writer>& w) {
-      w->finalize();
-      return snapshot_t(*w->storage);
-   }
-
-   static auto get_reader( const snapshot_t& buffer) {
-      return std::make_shared<reader>(buffer);
-   }
-
-   static snapshot_t load_from_file(const std::string& filename) {
-      snapshot_input_file<snapshot::json> file(filename);
-      return file.read();
-   }
-
-   static void write_to_file( const std::string& basename, const snapshot_t& snapshot ) {
-     snapshot_output_file<snapshot::json> file(basename);
-     file.write<snapshot_t>(snapshot);
-   }
-};
-
-struct buffered_snapshot_suite {
-   using writer_t = ostream_snapshot_writer;
-   using reader_t = istream_snapshot_reader;
-   using write_storage_t = std::ostringstream;
-   using snapshot_t = std::string;
-   using read_storage_t = std::istringstream;
-
-   struct writer : public writer_t {
-      writer( const std::shared_ptr<write_storage_t>& storage )
-      :writer_t(*storage)
-      ,storage(storage)
-      {
-
-      }
-
-      std::shared_ptr<write_storage_t> storage;
-   };
-
-   struct reader : public reader_t {
-      explicit reader(const std::shared_ptr<read_storage_t>& storage)
-      :reader_t(*storage)
-      ,storage(storage)
-      {}
-
-      std::shared_ptr<read_storage_t> storage;
-   };
-
-
-   static auto get_writer() {
-      return std::make_shared<writer>(std::make_shared<write_storage_t>());
-   }
-
-   static auto finalize(const std::shared_ptr<writer>& w) {
-      w->finalize();
-      return w->storage->str();
-   }
-
-   static auto get_reader( const snapshot_t& buffer) {
-      return std::make_shared<reader>(std::make_shared<read_storage_t>(buffer));
-   }
-
-   static snapshot_t load_from_file(const std::string& filename) {
-      snapshot_input_file<snapshot::binary> file(filename);
-      return file.read_as_string();
-   }
-
-   static void write_to_file( const std::string& basename, const snapshot_t& snapshot ) {
-      snapshot_output_file<snapshot::binary> file(basename);
-      file.write<snapshot_t>(snapshot);
-   }
-};
-
 BOOST_AUTO_TEST_SUITE(snapshot_tests)
-
-using snapshot_suites = boost::mpl::list<variant_snapshot_suite, buffered_snapshot_suite>;
 
 namespace {
    void variant_diff_helper(const fc::variant& lhs, const fc::variant& rhs, std::function<void(const std::string&, const fc::variant&, const fc::variant&)>&& out){
@@ -278,8 +175,6 @@ namespace {
       }
       BOOST_REQUIRE_EQUAL(lhs_integrity_hash.str(), rhs_integrity_hash.str());
    }
-
-   eosio::producer_plugin::snapshot_information test_snap_info;
 }
 
 template<typename SnapshotSuite>
@@ -810,53 +705,6 @@ BOOST_AUTO_TEST_CASE_TEMPLATE(test_kv_snapshot, SNAPSHOT_SUITE, snapshot_suites)
          }
       }
    }
-}
-
-using next_t = eosio::producer_plugin::next_function<eosio::producer_plugin::snapshot_information>;
-
-BOOST_AUTO_TEST_CASE_TEMPLATE(test_snapshot_information, SNAPSHOT_SUITE, snapshot_suites) {
-   tester chain;
-   const chainbase::bfs::path parent_path = chain.get_config().blog.log_dir.parent_path();
-
-   chain.create_account("snapshot"_n);
-   chain.produce_blocks(1);
-   chain.set_code("snapshot"_n, contracts::snapshot_test_wasm());
-   chain.set_abi("snapshot"_n, contracts::snapshot_test_abi().data());
-   chain.produce_blocks(1);
-
-   auto block = chain.produce_block();
-   BOOST_REQUIRE_EQUAL(block->block_num(), 6); // ensure that test setup stays consistent with original snapshot setup
-   // undo the auto-pending from tester
-   chain.control->abort_block();
-
-   auto block2 = chain.produce_block();
-   BOOST_REQUIRE_EQUAL(block2->block_num(), 7); // ensure that test setup stays consistent with original snapshot setup
-   // undo the auto-pending from tester
-   chain.control->abort_block();
-
-   // write snapshot
-   auto write_snapshot = [&]( const bfs::path& p ) -> void {
-      if ( !bfs::exists( p.parent_path() ) )
-         bfs::create_directory( p.parent_path() );
-
-      // create the snapshot
-      auto snap_out = std::ofstream(p.generic_string(), (std::ios::out | std::ios::binary));
-      auto writer = std::make_shared<ostream_snapshot_writer>(snap_out);
-      (*chain.control).write_snapshot(writer);
-      writer->finalize();
-      snap_out.flush();
-      snap_out.close();
-   };
-
-   auto final_path = eosio::pending_snapshot::get_final_path(block2->previous, "../snapshots/");
-   auto pending_path = eosio::pending_snapshot::get_pending_path(block2->previous, "../snapshots/");
-
-   write_snapshot( pending_path );
-   next_t next;
-   eosio::pending_snapshot pending{ block2->previous, next, pending_path.generic_string(), final_path.generic_string() };
-   test_snap_info = pending.finalize(*chain.control);
-   BOOST_REQUIRE_EQUAL(test_snap_info.head_block_num, 6);
-   BOOST_REQUIRE_EQUAL(test_snap_info.version, chain_snapshot_header::current_version);
 }
 
 BOOST_AUTO_TEST_SUITE_END()
